Extract cli into its own module qpc to resolve install issue. Closes #82
* Updates Makefile test, and test-coverage targets to capture all relocation changes.
