+++ qpc/auth/tests_clear.py
@@ -16,10 +16,10 @@ from io import StringIO
 from argparse import ArgumentParser, Namespace
 import requests
 import requests_mock
-from cli.tests_utilities import HushUpStderr, redirect_stdout
-from cli.request import BASE_URL, CONNECTION_ERROR_MSG, SSL_ERROR_MSG
-from cli.auth import AUTH_URI
-from cli.auth.clear import AuthClearCommand
+from qpc.tests_utilities import HushUpStderr, redirect_stdout
+from qpc.request import BASE_URL, CONNECTION_ERROR_MSG, SSL_ERROR_MSG
+from qpc.auth import AUTH_URI
+from qpc.auth.clear import AuthClearCommand
 
 PARSER = ArgumentParser()
 SUBPARSER = PARSER.add_subparsers(dest='subcommand')
