+++ qpc/utils.py
